+++ src/typing_ex/handler/core.clj
@@ -13,6 +13,8 @@
    [taoensso.timbre :as timbre :refer [debug]]
    [ring.util.anti-forgery :refer [anti-forgery-field]]))
 
+(timbre/set-level! :debug)
+
 (def DAYS 30)
 
 ;; FIXME: データベースに持っていこ。
@@ -64,9 +66,7 @@
 
 ;; index.
 (defmethod ig/init-key :typing-ex.handler.core/typing [_ _]
-  (fn [{token :anti-forgery-token :as req}]
-    ;; FIXME これを利用できないか？
-    (debug "/typing (:anti-forgery-token req):" token)
+  (fn [_]
     [::response/ok
      (str
       "<!DOCTYPE html>
@@ -93,12 +93,15 @@
 ;; FIXME: CSRF post
 (defmethod ig/init-key :typing-ex.handler.core/score-post [_ {:keys [db]}]
   (fn [req]
-    (let [pt (get-in req [:params :pt])
-          nick (get-nick req)
-          rcv {:pt (Integer/parseInt pt) :users_nick nick}]
-      (debug "/score-post pt" rcv)
-      (results/insert-pt db rcv)
-      [::response/ok (str rcv)])))
+    (timbre/debug "score-post:" (str req))
+    [::response/ok (str {:pt 100 :users_nick "you"})]))
+  ;; (fn [req]
+  ;;   (let [pt (get-in req [:params :pt])
+  ;;         nick (get-nick req)
+  ;;         rcv {:pt (Integer/parseInt pt) :users_nick nick}]
+  ;;     (debug "/score-post pt" rcv)
+  ;;     (results/insert-pt db rcv)
+  ;;     [::response/ok (str rcv)])))
 
 (defmethod ig/init-key :typing-ex.handler.core/scores [_ {:keys [db]}]
   (fn [req]