+++ CHANGELOG.md
@@ -10,17 +10,18 @@
 ## Unreleased
 - 各行ごとに集計
 - 認証を l22 データベースで
-- nick コラムは login コラムに名称替え。
+- コラム名 nick を login に名称替えをする。
 - users_nick がよろしくない。login だな。新しくは。
 - ページのデザイン。
   - タイプし終えたら自動でボーナス。
   - detail よりも、回数、平均、最高点を表示したら？
-
+- app.melt へのデプロイ。
 
 ## 1.2.0-SNAPSHOT
 ### Changed
 - typing.cljs: scores ボタンを logout ボタンの左に。
 - タイプを始めると自動でスタートする。
+- boudary/users.clj: ds をやめ、全て ds-opt を使うようにした。
 
 ## 1.1.0 - 2022-03-14
 ### Changed